refactor(mapgen): centralize typed-array admission
