+++ packages/mapgen-core/test/authoring/schema/typed-array.test.ts
@@ -6,6 +6,11 @@ import {
   TypedArraySchemas,
   typedArrayConstructorFor,
 } from "@mapgen/authoring/schema/typed-array.js";
+import {
+  compileTypedArrayAdmissionPlan,
+  validateTypedArrayAdmission,
+} from "@mapgen/authoring/schema/typed-array-admission.js";
+import { Type } from "typebox";
 
 function runtimeMetadata(schema: object): unknown {
   return (schema as Record<PropertyKey, unknown>)["x-runtime"];
@@ -32,12 +37,95 @@ describe("typed-array schemas and guards", () => {
       ctor: "Float32Array",
       cardinality: "constructor-only",
     });
+    const mapGrid = TypedArraySchemas.u16({ cardinality: "map-grid" });
+    expect(runtimeMetadata(mapGrid)).toEqual({
+      kind: "typed-array",
+      ctor: "Uint16Array",
+      cardinality: "map-grid",
+    });
+    expect(Object.getOwnPropertyDescriptor(mapGrid, "x-runtime")?.enumerable).toBe(true);
+    expect(JSON.parse(JSON.stringify(mapGrid))["x-runtime"]).toEqual({
+      kind: "typed-array",
+      ctor: "Uint16Array",
+      cardinality: "map-grid",
+    });
   });
 
   it("uses one exact constructor registry for schema admission", () => {
     expect(isSupportedTypedArrayName("Uint16Array")).toBe(true);
     expect(isSupportedTypedArrayName("Float64Array")).toBe(false);
     expect(typedArrayConstructorFor("Uint16Array")).toBe(Uint16Array);
+  });
+
+  it("admits map-grid only from supplied dimensions and fails closed without them", () => {
+    const schema = Type.Object({
+      values: TypedArraySchemas.u8({ cardinality: "map-grid" }),
+    });
+    const plan = compileTypedArrayAdmissionPlan(schema, {
+      subject: "Test",
+      contextualCardinality: "allow",
+    });
+
+    expect(validateTypedArrayAdmission(plan, { values: new Uint8Array(6) })).toEqual([
+      {
+        code: "typed-array-cardinality-source",
+        path: "$.values",
+        sourcePath: "map-grid",
+        observed: undefined,
+      },
+    ]);
+    expect(
+      validateTypedArrayAdmission(
+        plan,
+        { values: new Uint8Array(6) },
+        { dimensions: { width: 2, height: 3 } }
+      )
+    ).toEqual([]);
+    expect(
+      validateTypedArrayAdmission(
+        plan,
+        { values: new Uint8Array(5) },
+        { dimensions: { width: 2, height: 3 } }
+      )
+    ).toEqual([
+      {
+        code: "typed-array-cardinality",
+        path: "$.values",
+        cardinalityPaths: ["map-grid"],
+        addend: 0,
+        expectedLength: 6,
+        observedLength: 5,
+      },
+    ]);
+    expect(
+      validateTypedArrayAdmission(
+        plan,
+        { values: new Uint8Array(0) },
+        { dimensions: { width: 2.5, height: 3 } }
+      )
+    ).toEqual([
+      {
+        code: "typed-array-cardinality-source",
+        path: "$.values",
+        sourcePath: "map-grid.width",
+        observed: 2.5,
+      },
+    ]);
+    expect(
+      validateTypedArrayAdmission(
+        plan,
+        { values: new Uint8Array(0) },
+        { dimensions: { width: Number.MAX_SAFE_INTEGER, height: 2 } }
+      )
+    ).toEqual([
+      {
+        code: "typed-array-cardinality-overflow",
+        path: "$.values",
+        cardinalityPaths: ["map-grid"],
+        factors: [Number.MAX_SAFE_INTEGER, 2],
+        addend: 0,
+      },
+    ]);
   });
 
   it("refuses wrong constructors, subclasses, spoofed prototypes, and non-array views", () => {
